- New skeletons for the updated version of Legions. - Bug fixes.
git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@970 0874c2aa-48f6-4a0f-8a6b-e32f4465a257

+++ src/android2iphone/android/graphics/Canvas.java
@@ -141,7 +141,7 @@ public class Canvas {
         releaseCGContext();
     }
 
-    public void drawText(String texttodisplay, int left, int top, Paint paint) {
+    public void drawText(String texttodisplay, float left, float top, Paint paint) {
         createCGContext();
         context.setFillColor(paint.xmlvmGetColor());
         NSString.drawAtPoint(texttodisplay, new CGPoint(left, top), paint.xmlvmGetUIFont());
